migrate logic out of metadata agent
